# deliverables/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Sentinel AML \u2014 Transaction Monitoring System",
        "description": "Real-time money-laundering detection for MeridianTrust Bank.\n\n**Flow:** ingestion \u2192 detection \u2192 risk-scored alert \u2192 case disposition.\n\n**Roles** (RBAC is enforced at this API layer, not only in the UI):\n- `ANALYST` \u2014 alert queue (PII masked), cases, dispositions\n- `SENIOR_ANALYST` \u2014 the above, plus unmasked PII and SAR escalation\n- `COMPLIANCE_ADMIN` \u2014 the above, plus ingestion and rule configuration\n\nDetection thresholds are tunable at runtime via\n`PATCH /api/v1/admin/rules/{ruleCode}` \u2014 no redeployment required.\n",
        "contact": {
            "name": "MeridianTrust Financial Crime Engineering"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "security": [
        {
            "basicAuth": []
        }
    ],
...
    "paths": {
        "/api/v1/admin/fx-rates/{currency}": {
            "put": {
                "tags": [
                    "Administration"
                ],
                "summary": "Set an exchange rate",
                "description": "Applies to transactions ingested from this point on; historical amounts keep the rate they were normalised with.",
                "operationId": "updateFxRate",
                "parameters": [
                    {
                        "name": "currency",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FxRateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FxRate"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/watchlist": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Named counterparty watchlist",
                "operationId": "watchlist",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WatchlistCounterparty"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Administration"
                ],
                "summary": "Add a counterparty to the watchlist",
                "operationId": "addWatchlist",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WatchlistCounterparty"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WatchlistCounterparty"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/jurisdictions": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "High-risk and sanctioned jurisdiction list",
                "operationId": "jurisdictions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HighRiskJurisdiction"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Administration"
                ],
                "summary": "Add or update a jurisdiction listing",
                "operationId": "upsertJurisdiction",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HighRiskJurisdiction"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HighRiskJurisdiction"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/rules/{ruleCode}": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Get one rule's configuration",
                "operationId": "rule",
                "parameters": [
                    {
                        "name": "ruleCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuleDto"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Administration"
                ],
                "summary": "Tune a rule at runtime",
                "description": "Changes take effect on the next detection run \u2014 no restart, no\nredeployment. Parameters are validated before they are accepted,\nso a configuration that could never fire is rejected rather than\nsilently disabling a control.\n\nExample \u2014 loosen structuring to catch pairs rather than triples:\n`{\"params\": {\"minCount\": 2, \"windowHours\": 24,\n             \"lowerBound\": 9000, \"upperBound\": 9999.99}}`\n",
                "operationId": "updateRule",
                "parameters": [
                    {
                        "name": "ruleCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RuleUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Rule updated and cache refreshed",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuleDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parameters failed validation",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuleDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Requires COMPLIANCE_ADMIN",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuleDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No such rule",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RuleDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/jurisdictions/{countryCode}": {
            "delete": {
                "tags": [
                    "Administration"
                ],
                "summary": "Deactivate a jurisdiction listing",
                "description": "Deactivates rather than deletes \u2014 historical alerts citing this listing must remain explicable.",
                "operationId": "deactivateJurisdiction",
                "parameters": [
                    {
                        "name": "countryCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "FxRateRequest": {
                "type": "object",
                "properties": {
                    "rateToBase": {
                        "type": "number",
                        "minimum": 1e-08
                    }
                },
                "required": [
                    "rateToBase"
                ]
            },
            "FxRate": {
                "type": "object",
                "properties": {
                    "currency": {
                        "type": "string"
                    },
                    "rateToBase": {
                        "type": "number"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "WatchlistCounterparty": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "normalizedName": {
                        "type": "string"
                    },
                    "listType": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                }
            },
            "HighRiskJurisdiction": {
                "type": "object",
                "properties": {
                    "countryCode": {
                        "type": "string"
                    },
                    "countryName": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "riskWeight": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "source": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                }
            },
...
            "RuleUpdateRequest": {
                "type": "object",
                "description": "Partial rule update \u2014 omitted fields are left unchanged.",
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    },
                    "weight": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Base contribution to the risk score, 0-100"
                    },
                    "severity": {
                        "type": "string",
                        "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH",
                            "CRITICAL"
                        ]
                    },
                    "params": {
                        "type": "object",
                        "additionalProperties": {},
                        "description": "Rule-specific thresholds and time windows"
                    }
                }
            },
            "RuleDto": {
                "type": "object",
                "properties": {
                    "ruleCode": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "typology": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "weight": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "severity": {
                        "type": "string"
                    },
                    "params": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "updatedBy": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "description": "HTTP Basic. Demo users: analyst / senior / admin.",
                "scheme": "basic"
            }
        }
    }
}